update_vpn_ddns/src/update_vpn_ddns/update_vpn_ddns.py
import os
import gspread
import datetime
import requests

from dotenv import load_dotenv
from oauth2client.service_account import ServiceAccountCredentials


def update_dns_record(...):
    # DNS update logic here
    pass

def upload_ip(...):
    # Google Sheets upload for IP
    pass

def main():
    load_dotenv()
    # Function calls here
    update_dns_record(...)
    upload_ip(...)
