update_vpn_ddns/requirements.txt
# Google Services API libraries
gspread  # Used for interacting with Google Sheets
oauth2client  # Google authentication with service accounts

# HTTP requests
requests  # For making external API endpoint requests

# Environment variables
python-dotenv  # Loads .env files for secrets and configs

# Date and time utilities (not needed for requirements.txt since built-in)
# datetime
# os
# logging